extract idempotency history lookup route
